requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
pydantic>=2.6.0
pyyaml>=6.0.1
weasyprint>=61.0
python-multipart>=0.0.9
requests>=2.31.0
